Remove Spark(Streaming)PipelineOptionsFactory
Pipeline authors should generally not use any runner-specific classes,
but instead should select the runner and appropriate configurations
through the PipelineOptionsFactory.fromArgs() method. The runner can
then obtain the appropriately typed PipelineOptions class as required
and do any neccessary validation. Failing this, they should use the
provided PipelineOptions#as() method to acquire the appropriately typed
options.

If required, users should construct SparkPipelineOptions via
PipelineOptionsFactory.as(SparkPipelineOptions.class).
