+++ runners/spark/src/test/java/org/apache/beam/runners/spark/coders/WritableCoderTest.java
@@ -19,6 +19,7 @@
 package org.apache.beam.runners.spark.coders;
 
 import org.apache.beam.sdk.testing.CoderProperties;
+
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.NullWritable;
 import org.junit.Test;
